feat: brought it to a cleaner state
