+++ lib/controllers/single_room_controller.dart
@@ -12,7 +12,8 @@ import 'package:resonate/models/participant.dart';
 import 'package:resonate/routes/app_routes.dart';
 import 'package:resonate/services/appwrite_service.dart';
 import 'package:resonate/services/room_service.dart';
-import 'package:resonate/views/widgets/loading_widget.dart';
+import 'package:resonate/views/widgets/loading_dialog.dart';
+
 
 import '../utils/constants.dart';
 
@@ -188,7 +189,7 @@ class SingleRoomController extends GetxController {
   }
 
   Future<void> leaveRoom() async {
-    loadingWidget();
+    loadingDialog(Get.context!);
     await RoomService.leaveRoom(roomId: appwriteRoom.id);
     Get.delete<SingleRoomController>();
   }
