+++ lib/main.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 import 'package:resonate/firebase_options.dart';
-import 'package:resonate/new_themes/new_theme.dart';
+import 'package:resonate/themes/new_theme.dart';
 import 'package:resonate/routes/app_pages.dart';
 import 'package:resonate/routes/app_routes.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:resonate/new_themes/theme_list.dart';
+import 'package:resonate/themes/theme_list.dart';
 import 'package:resonate/utils/ui_sizes.dart';
-import 'new_themes/new_theme_screen_controller.dart';
+import 'themes/new_theme_screen_controller.dart';
 import 'themes/theme_controller.dart';
 
 Future<void> main() async {
@@ -32,11 +32,8 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     UiSizes.init(context);
-
-    // This line is just for initialization of theme controller. This will prevent some breaking changes in old ui, until it gets fixed
-    Get.put(ThemeController());
-
     final newThemeController = Get.put(NewThemeController());
+    Get.put(ThemeController());
 
     return Obx(
       () => GetMaterialApp(
@@ -59,19 +56,5 @@ class MyApp extends StatelessWidget {
         getPages: AppPages.pages,
       ),
     );
-
-    // return Obx(
-    //   () => GetMaterialApp(
-    //     debugShowCheckedModeBanner: false,
-    //     title: 'Resonate',
-    //     theme: NewTheme.classicLightTheme,
-    //     // theme: Themes.getLightTheme(themeController.primaryColor.value),
-    //     // darkTheme: Themes.getDarkTheme(themeController.primaryColor.value),
-    //     darkTheme: Themes.getDarkTheme(themeController.primaryColor.value),
-    //     themeMode: themeController.theme,
-    //     initialRoute: AppRoutes.splash,
-    //     getPages: AppPages.pages,
-    //   ),
-    // );
   }
 }
