+++ lib/themes/theme_controller.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:resonate/models/themes_model.dart';
 
 import '../utils/enums/themes_enum.dart';
