+++ lib/views/screens/email_verification_screen.dart
@@ -1,445 +1,259 @@
-import 'dart:ui';
 import 'package:circular_countdown_timer/circular_countdown_timer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/semantics.dart';
 import 'package:flutter_otp_text_field/flutter_otp_text_field.dart';
 import 'package:get/get.dart';
-import 'package:loading_animation_widget/loading_animation_widget.dart';
-import 'package:resonate/controllers/authentication_controller.dart';
-import 'package:resonate/routes/app_routes.dart';
-import 'package:resonate/themes/theme_controller.dart';
-import 'package:resonate/utils/colors.dart';
-import 'package:resonate/utils/enums/message_type_enum.dart';
-import 'package:resonate/utils/ui_sizes.dart';
-import 'package:resonate/views/widgets/snackbar.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:resonate/views/widgets/loading_dialog.dart';
 
+import '../../controllers/authentication_controller.dart';
 import '../../controllers/email_verify_controller.dart';
+import '../../routes/app_routes.dart';
+import '../../utils/enums/message_type_enum.dart';
+import '../../utils/ui_sizes.dart';
+import '../widgets/snackbar.dart';
 
 class EmailVerificationScreen extends StatelessWidget {
-  EmailVerificationScreen({Key? key}) : super(key: key);
+  EmailVerificationScreen({super.key});
 
   final controller = Get.find<AuthenticationController>();
   final emailVerifyController = Get.find<EmailVerifyController>();
 
-  final ThemeController themeController = Get.find<ThemeController>();
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: SafeArea(
-      child: Obx(
-        () => Stack(children: [
-          !controller.authStateController.isInitializing.value
-              ? SingleChildScrollView(
-                  child: Container(
-                    padding: EdgeInsets.symmetric(
-                        horizontal: UiSizes.width_20,
-                        vertical: UiSizes.height_10),
-                    child: Column(
-                      children: [
-                        Align(
-                          alignment: Alignment.centerLeft,
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Text(
-                                "Resonate",
-                                style: TextStyle(
-                                    color: themeController.primaryColor.value,
-                                    fontSize: UiSizes.size_25),
-                              ),
-                              ClipRRect(
-                                borderRadius: BorderRadius.circular(100),
-                                child: Material(
-                                  child: InkWell(
-                                    onTap: () =>
-                                        {Get.toNamed(AppRoutes.tabview)},
-                                    child: Ink(
-                                        height: UiSizes.height_30,
-                                        width: UiSizes.width_100,
-                                        decoration: BoxDecoration(
-                                          gradient: themeController
-                                              .createDynamicGradient(),
-                                          borderRadius:
-                                              BorderRadius.circular(20),
-                                        ),
-                                        child: Center(
-                                          child: Text("Back",
-                                              style: TextStyle(
-                                                  color: Theme.of(context)
-                                                      .colorScheme
-                                                      .surface,
-                                                  fontSize: UiSizes.size_14)),
-                                        )),
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                        SizedBox(height: UiSizes.height_80),
-                        Align(
-                          alignment: Alignment.topCenter,
-                          child: Container(
-                            height: UiSizes.height_40,
-                            decoration: const BoxDecoration(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(40)),
-                            ),
-                            child: Center(
-                                child: Text(
-                              "OTP Verification",
-                              style: TextStyle(fontSize: UiSizes.size_24),
-                            )),
-                          ),
-                        ),
-                        SizedBox(height: UiSizes.height_90),
-                        Stack(children: [
-                          Column(
-                            children: [
-                              SizedBox(
-                                height: UiSizes.height_15,
-                              ),
-                              Container(
-                                decoration: BoxDecoration(
-                                  borderRadius: const BorderRadius.all(
-                                      Radius.circular(20)),
-                                  color: Theme.of(context)
-                                      .colorScheme
-                                      .primaryContainer
-                                      .withOpacity(.15),
-                                ),
-                                padding: EdgeInsets.symmetric(
-                                  horizontal: UiSizes.width_16,
-                                  vertical: UiSizes.height_30,
-                                ),
-                                child: Column(
-                                  children: [
-                                    SizedBox(
-                                      height: UiSizes.height_10,
-                                    ),
-                                    Text(
-                                      'Enter the OTP sent to ${controller.authStateController.email}',
-                                      textAlign: TextAlign.center,
-                                      style: TextStyle(
-                                        fontSize: UiSizes.size_14,
-                                      ),
-                                    ),
-                                    SizedBox(height: UiSizes.height_40),
-                                    OtpTextField(
-                                      contentPadding: EdgeInsets.zero,
-                                      textStyle: TextStyle(
-                                          fontSize: UiSizes.size_14,
-                                          color: Colors.red),
-                                      fieldWidth: UiSizes.width_40,
-                                      numberOfFields: 6,
-                                      focusedBorderColor: const Color.fromARGB(
-                                          224, 68, 170, 50),
-                                      borderWidth: UiSizes.width_1_5,
-                                      clearText: emailVerifyController
-                                          .clearTextField.value,
-                                      enabledBorderColor:
-                                          themeController.primaryColor.value,
-                                      showFieldAsBox: true,
-                                      //runs when every text-field is filled
-                                      onSubmit:
-                                          (String verificationCode) async {
-                                        emailVerifyController
-                                            .isVerifying.value = true;
-                                        await emailVerifyController
-                                            .verifyOTP(verificationCode);
-                                        if (emailVerifyController
-                                                .responseVerify.responseBody ==
-                                            '{"message":"null"}') {
-                                          String result =
-                                              await emailVerifyController
-                                                  .checkVerificationStatus();
-                                          if (result == "true") {
-                                            customSnackbar(
-                                              "Verification Complete",
-                                              "Congratulations you have verified your Email",
-                                              MessageType.success,
-                                            );
-
-                                            SemanticsService.announce(
-                                              "Congratulations you have verified your Email",
-                                              TextDirection.ltr,
-                                            );
-                                            await emailVerifyController
-                                                .setVerified();
-                                            if (emailVerifyController
-                                                    .responseSetVerified
-                                                    .responseBody ==
-                                                '{"message":"null"}') {
-                                              emailVerifyController
-                                                  .isVerifying.value = false;
-                                              controller.authStateController
-                                                  .setUserProfileData();
-                                              Get.offAllNamed(
-                                                  AppRoutes.tabview);
-                                            } else {
-                                              emailVerifyController
-                                                  .isVerifying.value = false;
-                                              customSnackbar(
-                                                'Oops',
-                                                emailVerifyController
-                                                    .responseSetVerified
-                                                    .responseBody,
-                                                MessageType.error,
-                                              );
-
-                                              SemanticsService.announce(
-                                                emailVerifyController
-                                                    .responseSetVerified
-                                                    .responseBody,
-                                                TextDirection.ltr,
-                                              );
-                                            }
-                                          } else {
-                                            emailVerifyController
-                                                .isVerifying.value = false;
-                                            customSnackbar(
-                                              "Verification Failed",
-                                              "OTP mismatch occurred please try again",
-                                              MessageType.error,
-                                            );
-
-                                            SemanticsService.announce(
-                                              "OTP mismatch occurred please try again",
-                                              TextDirection.ltr,
-                                            );
-                                          }
-                                        } else {
-                                          customSnackbar(
-                                            'Oops',
-                                            emailVerifyController
-                                                .responseVerify.responseBody,
-                                            MessageType.error,
-                                          );
-
-                                          SemanticsService.announce(
-                                            emailVerifyController
-                                                .responseVerify.responseBody,
-                                            TextDirection.ltr,
-                                          );
-                                        }
-                                      },
-                                    ),
-                                    SizedBox(height: UiSizes.height_60),
-                                    Obx(
-                                      () => Container(
-                                        width: UiSizes.width_300,
-                                        decoration: BoxDecoration(
-                                            border: Border.all(
-                                                color: themeController
-                                                    .primaryColor.value
-                                                    .withOpacity(0.6),
-                                                width: UiSizes.width_3),
-                                            borderRadius:
-                                                BorderRadius.circular(70),
-                                            gradient: emailVerifyController
-                                                    .resendIsAllowed.value
-                                                ? themeController
-                                                    .createDynamicGradient()
-                                                : const LinearGradient(colors: [
-                                                    Color.fromARGB(
-                                                        255, 74, 74, 74),
-                                                    Color.fromARGB(
-                                                        255, 68, 63, 63)
-                                                  ])),
-                                        child: SizedBox(
-                                          height: UiSizes.height_45,
-                                          width: UiSizes.width_294,
-                                          child: ElevatedButton(
-                                              style: ElevatedButton.styleFrom(
-                                                  shape: RoundedRectangleBorder(
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                              UiSizes.size_70),
-                                                      side: const BorderSide(
-                                                          width: 0,
-                                                          color: Colors
-                                                              .transparent)),
-                                                  backgroundColor:
-                                                      Colors.transparent,
-                                                  shadowColor:
-                                                      Colors.transparent),
-                                              onPressed: emailVerifyController
-                                                      .resendIsAllowed.value
-                                                  ? () {
-                                                      emailVerifyController
-                                                          .resendIsAllowed
-                                                          .value = false;
-                                                      emailVerifyController
-                                                          .clearTextField
-                                                          .value = true;
-                                                      emailVerifyController
-                                                          .sendOTP();
-                                                      customSnackbar(
-                                                        "OTP Resent",
-                                                        "Please do check your mail for a new OTP",
-                                                        MessageType.info,
-                                                      );
-
-                                                      SemanticsService.announce(
-                                                        "Please do check your mail for a new OTP",
-                                                        TextDirection.ltr,
-                                                      );
-                                                    }
-                                                  : () {
-                                                      customSnackbar(
-                                                        "Hold on",
-                                                        "Please wait till the timer completes",
-                                                        MessageType.warning,
-                                                      );
-
-                                                      SemanticsService.announce(
-                                                        "Please wait till the timer completes",
-                                                        TextDirection.ltr,
-                                                      );
-                                                    },
-                                              child: emailVerifyController
-                                                      .resendIsAllowed.value
-                                                  ? Text(
-                                                      "Resend OTP",
-                                                      style: TextStyle(
-                                                        color: Colors.black,
-                                                        fontSize:
-                                                            UiSizes.size_14,
-                                                      ),
-                                                    )
-                                                  : Row(
-                                                      mainAxisAlignment:
-                                                          MainAxisAlignment
-                                                              .start,
-                                                      children: [
-                                                        Text(
-                                                          "Resend OTP?",
-                                                          style: TextStyle(
-                                                              fontSize: UiSizes
-                                                                  .size_14,
-                                                              color:
-                                                                  Colors.white),
-                                                        ),
-                                                        SizedBox(
-                                                          width:
-                                                              UiSizes.width_111,
-                                                        ),
-                                                        CircularCountDownTimer(
-                                                          textStyle: TextStyle(
-                                                              fontSize: UiSizes
-                                                                  .size_12,
-                                                              color: AppColor
-                                                                  .bgBlackColor),
-                                                          isTimerTextShown:
-                                                              true,
-                                                          isReverse: true,
-                                                          onComplete: () => {
-                                                            emailVerifyController
-                                                                .resendIsAllowed
-                                                                .value = true
-                                                          },
-                                                          width:
-                                                              UiSizes.width_33,
-                                                          height:
-                                                              UiSizes.height_33,
-                                                          duration: 30,
-                                                          backgroundColor:
-                                                              themeController
-                                                                  .primaryColor
-                                                                  .value,
-                                                          fillColor: const Color
-                                                              .fromARGB(
-                                                              255, 74, 74, 74),
-                                                          ringColor: const Color
-                                                              .fromARGB(255,
-                                                              130, 130, 130),
-                                                        ),
-                                                      ],
-                                                    )),
-                                        ),
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ],
-                          ),
-                          Align(
-                            alignment: Alignment.topCenter,
-                            child: GestureDetector(
-                                onTap: () {
-                                  emailVerifyController.isExpanded.value =
-                                      !emailVerifyController.isExpanded.value;
-                                  if (emailVerifyController.isExpanded.value ==
-                                      false) {
-                                    emailVerifyController.shouldDisplay.value =
-                                        false;
-                                  }
-                                },
-                                child: AnimatedContainer(
-                                    decoration: BoxDecoration(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .surfaceTint,
-                                      borderRadius: BorderRadius.circular(100),
-                                    ),
-                                    onEnd: () {
-                                      if (emailVerifyController
-                                              .isExpanded.value ==
-                                          true) {
-                                        emailVerifyController
-                                            .shouldDisplay.value = true;
-                                      }
-                                    },
-                                    height: UiSizes.height_35,
-                                    width:
-                                        emailVerifyController.isExpanded.value
-                                            ? UiSizes.width_180
-                                            : UiSizes.width_35,
-                                    // decoration: BoxDecoration(
-                                    //   borderRadius: BorderRadius.circular(8.5714285714 * Get.pixelRatio),
-                                    //   color: Theme.of(context).colorScheme.onPrimary,
-                                    // ),
-                                    duration: const Duration(milliseconds: 200),
-                                    child: emailVerifyController
-                                            .shouldDisplay.value
-                                        ? Center(
-                                            child: Text(
-                                              "Verify Your Email",
-                                              style: TextStyle(
-                                                fontSize: UiSizes.size_14,
-                                                color: Theme.of(context)
-                                                    .colorScheme
-                                                    .surface,
-                                              ),
-                                            ),
-                                          )
-                                        : const SizedBox())),
-                          )
-                        ]),
-                      ],
+      appBar: AppBar(),
+      body: Container(
+        padding: EdgeInsets.symmetric(
+          horizontal: UiSizes.width_20,
+          vertical: UiSizes.height_20,
+        ),
+        width: double.maxFinite,
+        child: Form(
+          child: Column(
+            children: [
+              SizedBox(
+                height: UiSizes.height_60,
+              ),
+              MergeSemantics(
+                child: Column(
+                  children: [
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Text(
+                        "Enter your\nVerification Code",
+                        style: Theme.of(context).textTheme.headlineMedium,
+                      ),
                     ),
-                  ),
-                )
-              : BackdropFilter(
-                  filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
-                  child: Center(
-                      child: LoadingAnimationWidget.threeRotatingDots(
-                          color: themeController.primaryColor.value,
-                          size: Get.pixelRatio * 20)),
+                    SizedBox(
+                      height: UiSizes.height_40,
+                    ),
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: RichText(
+                        text: TextSpan(
+                          style: TextStyle(
+                            fontFamily: GoogleFonts.poppins().fontFamily,
+                            color:
+                                Theme.of(context).brightness == Brightness.light
+                                    ? Colors.black
+                                    : Colors.white,
+                          ),
+                          children: [
+                            const TextSpan(
+                              text: "We sent a 6-digit verification code to\n",
+                            ),
+                            TextSpan(
+                              text: controller.authStateController.email,
+                              style: TextStyle(
+                                fontWeight: FontWeight.bold,
+                                color: Theme.of(context).colorScheme.primary,
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                  ],
                 ),
-          emailVerifyController.isVerifying.value
-              ? BackdropFilter(
-                  filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
-                  child: Center(
-                      child: LoadingAnimationWidget.threeRotatingDots(
-                          color: themeController.primaryColor.value,
-                          size: Get.pixelRatio * 20)),
-                )
-              : const SizedBox(),
-        ]),
+              ),
+              SizedBox(
+                height: UiSizes.height_60,
+              ),
+              OtpTextField(
+                autoFocus: true,
+                numberOfFields: 6,
+                showFieldAsBox: true,
+                keyboardType: TextInputType.number,
+                filled: true,
+                fillColor: Theme.of(context).colorScheme.secondary,
+                borderWidth: 1,
+                contentPadding: EdgeInsets.zero,
+                borderColor: Colors.transparent,
+                enabledBorderColor: Colors.transparent,
+                focusedBorderColor: Theme.of(context).colorScheme.primary,
+                // runs when every text-field is filled
+                onSubmit: (String verificationCode) async {
+                  loadingDialog(context);
+                  emailVerifyController.isVerifying.value = true;
+                  await emailVerifyController.verifyOTP(verificationCode);
+
+                  if (emailVerifyController.responseVerify.responseBody ==
+                      '{"message":"null"}') {
+                    String result =
+                        await emailVerifyController.checkVerificationStatus();
+                    if (result == "true") {
+                      customSnackbar(
+                        "Verification Complete",
+                        "Congratulations you have verified your Email",
+                        MessageType.success,
+                      );
+
+                      SemanticsService.announce(
+                        "Congratulations you have verified your Email",
+                        TextDirection.ltr,
+                      );
+                      await emailVerifyController.setVerified();
+                      if (emailVerifyController
+                              .responseSetVerified.responseBody ==
+                          '{"message":"null"}') {
+                        emailVerifyController.isVerifying.value = false;
+
+                        // Close loading dialog
+                        Get.back();
+
+                        controller.authStateController.setUserProfileData();
+                        Get.offAllNamed(AppRoutes.tabview);
+                      } else {
+                        emailVerifyController.isVerifying.value = false;
+
+                        // Close loading dialog
+                        Get.back();
+
+                        customSnackbar(
+                          'Oops',
+                          emailVerifyController
+                              .responseSetVerified.responseBody,
+                          MessageType.error,
+                        );
+
+                        SemanticsService.announce(
+                          emailVerifyController
+                              .responseSetVerified.responseBody,
+                          TextDirection.ltr,
+                        );
+                      }
+                    } else {
+                      emailVerifyController.isVerifying.value = false;
+
+                      // Close loading dialog
+                      Get.back();
+
+                      customSnackbar(
+                        "Verification Failed",
+                        "OTP mismatch occurred please try again",
+                        MessageType.error,
+                      );
+
+                      SemanticsService.announce(
+                        "OTP mismatch occurred please try again",
+                        TextDirection.ltr,
+                      );
+                    }
+                  } else {
+                    customSnackbar(
+                      'Oops',
+                      emailVerifyController.responseVerify.responseBody,
+                      MessageType.error,
+                    );
+
+                    SemanticsService.announce(
+                      emailVerifyController.responseVerify.responseBody,
+                      TextDirection.ltr,
+                    );
+                  }
+                },
+              ),
+              SizedBox(
+                height: UiSizes.height_60,
+              ),
+              Obx(
+                () => (emailVerifyController.resendIsAllowed.value)
+                    ? GestureDetector(
+                        onTap: () {
+                          emailVerifyController.resendIsAllowed.value = false;
+                          emailVerifyController.sendOTP();
+                          customSnackbar(
+                            "OTP resent",
+                            "Please check your mail for a new OTP.",
+                            MessageType.info,
+                          );
+
+                          SemanticsService.announce(
+                            "Please check your mail for a new OTP.",
+                            TextDirection.ltr,
+                          );
+                        },
+                        child: Text(
+                          "Request a new code",
+                          style: TextStyle(
+                            color: Theme.of(context).colorScheme.primary,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      )
+                    : MergeSemantics(
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Text(
+                              "Request a new code in",
+                              style: TextStyle(
+                                color:
+                                    Theme.of(context).colorScheme.onSecondary,
+                              ),
+                            ),
+                            Padding(
+                              padding: EdgeInsets.symmetric(
+                                horizontal: UiSizes.width_10,
+                              ),
+                              child: CircularCountDownTimer(
+                                textStyle: TextStyle(
+                                  color:
+                                      Theme.of(context).colorScheme.onSecondary,
+                                ),
+                                isTimerTextShown: true,
+                                isReverse: true,
+                                onComplete: () {
+                                  emailVerifyController.resendIsAllowed.value =
+                                      true;
+                                },
+                                width: 30,
+                                height: 30,
+                                duration: 30,
+                                backgroundColor:
+                                    Theme.of(context).colorScheme.secondary,
+                                fillColor:
+                                    Theme.of(context).colorScheme.primary,
+                                ringColor:
+                                    Theme.of(context).colorScheme.onSecondary,
+                              ),
+                            ),
+                            Text(
+                              "seconds.",
+                              style: TextStyle(
+                                color:
+                                    Theme.of(context).colorScheme.onSecondary,
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+              ),
+            ],
+          ),
+        ),
       ),
-    ));
+    );
   }
 }
