+++ lib/views/screens/explore_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+
+// For the Explore page
+
+class ExploreScreen extends StatelessWidget {
+  const ExploreScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(child: Text("Work in Progress"),),
+    );
+  }
+}
