+++ lib/views/screens/profile_screen.dart
@@ -1,338 +1,391 @@
-import 'dart:developer';
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:get/get.dart';
-import 'package:loading_animation_widget/loading_animation_widget.dart';
-import 'package:resonate/controllers/auth_state_controller.dart';
-import 'package:resonate/themes/theme_controller.dart';
-import 'package:resonate/utils/colors.dart';
-import 'package:resonate/utils/ui_sizes.dart';
-import 'package:resonate/views/widgets/color_selection_widget.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:resonate/views/screens/story_screen.dart';
+import 'package:resonate/views/widgets/loading_dialog.dart';
 
+import '../../controllers/auth_state_controller.dart';
 import '../../controllers/email_verify_controller.dart';
+import '../../models/mock_models/stories_model.dart';
 import '../../routes/app_routes.dart';
-import '../../utils/constants.dart';
-import '../widgets/custom_card.dart';
+import '../../utils/ui_sizes.dart';
 
 class ProfileScreen extends StatelessWidget {
-  ProfileScreen({final Key? key}) : super(key: key);
+  ProfileScreen({super.key});
 
-  final emailVerifyController =
-      Get.put<EmailVerifyController>(EmailVerifyController());
-  final authStateController =
-      Get.put<AuthStateController>(AuthStateController());
+  final List<StoriesModel> yourStories = [
+    StoriesModel(
+      title: "Sony",
+      description:
+          "Discover a wide range of high-quality products from Sony and the technology behind them, get instant access to our store and Entertainment Network.",
+      imgUrl: "assets/mock/sony.jpeg",
+    ),
+    StoriesModel(
+      title: "Coffee Girl",
+      description: "Morning coffee is good for health or not.",
+      imgUrl: "assets/mock/coffee_girl.jpeg",
+    ),
+    StoriesModel(
+      title: "Japan",
+      description: "Japan is a country with very religious people.",
+      imgUrl: "assets/mock/japan.jpeg",
+    ),
+  ];
 
-  final ThemeController themeController = Get.find<ThemeController>();
+  final List<StoriesModel> continueWithStories = [
+    StoriesModel(
+      title: "Podcast",
+      description:
+          "The sun dipped below the horizon, casting a warm, golden hue over the tranquil landscape. Birds began their evening songs, filling the air with a gentle melody that blended perfectly with the rustling of the leaves in the cool breeze.",
+      imgUrl: "assets/mock/podcast.jpeg",
+    ),
+    StoriesModel(
+      title: "Coffee Latte",
+      description: "Cafe latte is good for health or not.",
+      imgUrl: "assets/mock/cafe_latte.jpeg",
+    ),
+  ];
 
   @override
   Widget build(BuildContext context) {
+    final emailVerifyController =
+        Get.put<EmailVerifyController>(EmailVerifyController());
+
     return GetBuilder<AuthStateController>(
       builder: (controller) => Scaffold(
         appBar: AppBar(
-          title: const Text(
-            "Profile",
-          ),
-          actions: [
-            Row(
-              children: [
-                InkWell(
-                    borderRadius: BorderRadius.circular(100),
-                    onTap: () async {
-                      await authStateController.logout(context);
-                    },
-                    child: Icon(
-                      Icons.logout_rounded,
-                      color: Theme.of(context).colorScheme.primary,
-                    )),
-                const SizedBox(
-                  width: 20,
-                ),
-              ],
-            )
-          ],
+          title: const Text("Profile"),
         ),
-        body: Obx(
-          () => Center(
-              child: Stack(children: [
-            controller.isInitializing.value
-                ? BackdropFilter(
-                    filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
-                    child: Center(
-                        child: LoadingAnimationWidget.threeRotatingDots(
-                            color: Colors.red, size: Get.pixelRatio * 20)),
-                  )
-                : SingleChildScrollView(
-                    child: Column(
-                      children: <Widget>[
-                        SizedBox(
-                          height: UiSizes.height_45,
-                        ),
-                        Container(
-                          width: UiSizes.width_180,
-                          height: UiSizes.height_180,
-                          decoration: BoxDecoration(
-                            border: Border.all(
-                                color: themeController.primaryColor.value,
-                                width: UiSizes.width_4),
-                            shape: BoxShape.circle,
-                            image: DecorationImage(
-                              fit: BoxFit.cover,
-                              image: NetworkImage(
-                                  controller.profileImageUrl ?? ''),
-                            ),
+        body: SingleChildScrollView(
+          child: Column(
+            children: [
+              Container(
+                padding: EdgeInsets.symmetric(
+                  vertical: UiSizes.height_10,
+                  horizontal: UiSizes.width_20,
+                ),
+                width: double.maxFinite,
+                child: Column(
+                  children: [
+                    SizedBox(
+                      width: double.maxFinite,
+                      child: Row(
+                        children: [
+                          CircleAvatar(
+                            backgroundColor:
+                                Theme.of(context).colorScheme.secondary,
+                            backgroundImage:
+                                NetworkImage(controller.profileImageUrl ?? ''),
+                            radius: UiSizes.width_66,
                           ),
-                          child: Stack(
-                            children: [
-                              Align(
-                                alignment: Alignment.bottomRight,
-                                child: GestureDetector(
-                                  onTap: () {
-                                    emailVerifyController.isExpanded.value =
-                                        !(emailVerifyController
-                                            .isExpanded.value);
-                                    if (emailVerifyController
-                                            .isExpanded.value ==
-                                        false) {
-                                      emailVerifyController
-                                          .shouldDisplay.value = false;
-                                    }
-                                  },
-                                  child: AnimatedContainer(
-                                      onEnd: () {
-                                        if (emailVerifyController
-                                                .isExpanded.value ==
-                                            true) {
-                                          emailVerifyController
-                                              .shouldDisplay.value = true;
-                                        }
-                                      },
-                                      width:
-                                          emailVerifyController.isExpanded.value
-                                              ? UiSizes.width_170
-                                              : UiSizes.width_35,
-                                      decoration: BoxDecoration(
-                                          borderRadius:
-                                              BorderRadius.circular(100),
-                                          color: Colors.white),
-                                      duration:
-                                          const Duration(milliseconds: 300),
-                                      child: controller.isEmailVerified!
-                                          ? Row(
-                                              mainAxisAlignment:
-                                                  emailVerifyController
-                                                          .isExpanded.value
-                                                      ? MainAxisAlignment.start
-                                                      : MainAxisAlignment
-                                                          .center,
-                                              children: [
-                                                Icon(
-                                                  Icons.verified_rounded,
-                                                  color: AppColor.greenColor,
-                                                  size: UiSizes.size_32,
-                                                ),
-                                                emailVerifyController
-                                                        .shouldDisplay.value
-                                                    ? SizedBox(
-                                                        width: UiSizes.width_5,
-                                                      )
-                                                    : const SizedBox(
-                                                        height: 0,
-                                                        width: 0,
-                                                      ),
-                                                emailVerifyController
-                                                        .shouldDisplay.value
-                                                    ? Text(
-                                                        "Email Verified",
-                                                        style: TextStyle(
-                                                          color: Colors.black,
-                                                          fontSize:
-                                                              UiSizes.size_15,
-                                                        ),
-                                                      )
-                                                    : const SizedBox(
-                                                        height: 0,
-                                                        width: 0,
-                                                      )
-                                              ],
-                                            )
-                                          : Row(
-                                              mainAxisAlignment:
-                                                  emailVerifyController
-                                                          .isExpanded.value
-                                                      ? MainAxisAlignment.start
-                                                      : MainAxisAlignment
-                                                          .center,
-                                              children: [
-                                                Expanded(
-                                                  child: Icon(
-                                                    Icons.cancel_rounded,
-                                                    color: Colors.red,
-                                                    size: UiSizes.size_35,
-                                                  ),
-                                                ),
-                                                emailVerifyController
-                                                        .shouldDisplay.value
-                                                    ? SizedBox(
-                                                        width: UiSizes.width_10,
-                                                      )
-                                                    : const SizedBox(
-                                                        height: 0,
-                                                        width: 0,
-                                                      ),
-                                                emailVerifyController
-                                                        .shouldDisplay.value
-                                                    ? Text(
-                                                        "Verify Email",
-                                                        style: TextStyle(
-                                                            fontSize:
-                                                                UiSizes.size_15,
-                                                            color:
-                                                                Colors.black),
-                                                      )
-                                                    : const SizedBox(
-                                                        height: 0,
-                                                        width: 0,
-                                                      )
-                                              ],
-                                            )),
-                                ),
-                              ),
-                            ],
+                          SizedBox(
+                            width: UiSizes.width_20,
                           ),
-                        ),
-                        SizedBox(height: UiSizes.height_20),
-                        SizedBox(
-                          width: UiSizes.width_320,
-                          height: UiSizes.height_55,
-                          child: FittedBox(
-                            fit: BoxFit.contain,
-                            child: Text(
-                              "@ ${controller.userName}",
-                              style: TextStyle(
-                                  fontSize: UiSizes.size_35,
-                                  color: themeController.primaryColor.value),
-                            ),
-                          ),
-                        ),
-                        SizedBox(
-                          width: UiSizes.width_320,
-                          height: UiSizes.height_40,
-                          child: FittedBox(
-                            fit: BoxFit.contain,
-                            child: Text(
-                              controller.displayName.toString(),
-                              style: TextStyle(fontSize: UiSizes.size_40),
-                            ),
-                          ),
-                        ),
-                        SizedBox(height: UiSizes.height_20),
-                        !(controller.isEmailVerified!)
-                            ? ClipRRect(
-                                borderRadius: BorderRadius.circular(20),
-                                child: Material(
-                                  child: InkWell(
-                                    highlightColor:
-                                        const Color.fromARGB(138, 33, 140, 14),
-                                    splashColor:
-                                        const Color.fromARGB(172, 43, 174, 20),
-                                    onTap: () => {
-                                      emailVerifyController.isSending.value =
-                                          true,
-                                      emailVerifyController.sendOTP()
-                                    },
-                                    child: Ink(
-                                        height: UiSizes.height_40,
-                                        width: UiSizes.width_140,
-                                        decoration: BoxDecoration(
-                                            color: themeController
-                                                .primaryColor.value,
-                                            borderRadius:
-                                                BorderRadius.circular(20),
-                                            border: Border.all(
-                                                color: themeController
-                                                    .primaryColor.value,
-                                                width: 3)),
-                                        child: Center(
-                                          child: Text(
-                                            "Verify Email",
-                                            style: TextStyle(
-                                                color: Colors.black,
-                                                fontSize: UiSizes.size_14,
-                                                fontWeight: FontWeight.bold),
+                          Expanded(
+                            child: MergeSemantics(
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  Text(
+                                    controller.displayName.toString(),
+                                    style: TextStyle(
+                                      fontSize: UiSizes.size_24,
+                                      fontWeight: FontWeight.bold,
+                                      overflow: TextOverflow.ellipsis,
+                                    ),
+                                  ),
+                                  Text(
+                                    "@${controller.userName}",
+                                    style: TextStyle(
+                                      fontSize: UiSizes.size_14,
+                                      overflow: TextOverflow.ellipsis,
+                                    ),
+                                  ),
+                                  (controller.isEmailVerified!)
+                                      ? const Padding(
+                                          padding: EdgeInsets.only(top: 10),
+                                          child: Row(
+                                            children: [
+                                              Icon(
+                                                Icons.verified_user_outlined,
+                                                color: Colors.green,
+                                              ),
+                                              SizedBox(
+                                                width: 5,
+                                              ),
+                                              Text(
+                                                "Verified user",
+                                                style: TextStyle(
+                                                    color: Colors.green),
+                                              ),
+                                            ],
                                           ),
-                                        )),
+                                        )
+                                      : const SizedBox(),
+                                ],
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                    (!controller.isEmailVerified!)
+                        ? Container(
+                            margin: EdgeInsets.only(
+                              top: UiSizes.height_10,
+                            ),
+                            width: double.maxFinite,
+                            child: OutlinedButton(
+                              onPressed: () {
+                                // Display Loading Dialog
+                                loadingDialog(context);
+                                emailVerifyController.isSending.value = true;
+                                emailVerifyController.sendOTP();
+                              },
+                              child: const Row(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Icon(Icons.verified_user_outlined),
+                                  SizedBox(
+                                    width: 10,
+                                  ),
+                                  Text("Verify Email"),
+                                ],
+                              ),
+                            ),
+                          )
+                        : const SizedBox(),
+                    SizedBox(
+                      height: UiSizes.height_10,
+                    ),
+                    SizedBox(
+                      width: double.maxFinite,
+                      child: Row(
+                        children: [
+                          Expanded(
+                            child: SizedBox(
+                              height: 34,
+                              child: ElevatedButton(
+                                onPressed: () {
+                                  Get.toNamed(AppRoutes.editProfile);
+                                },
+                                child: const FittedBox(
+                                  child: Padding(
+                                    padding: EdgeInsets.all(8.0),
+                                    child: Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      children: [
+                                        Icon(Icons.edit),
+                                        SizedBox(
+                                          width: 10,
+                                        ),
+                                        Text("Edit Profile"),
+                                      ],
+                                    ),
                                   ),
                                 ),
-                              )
-                            : const SizedBox(),
-                        SizedBox(height: UiSizes.height_20),
-                        ColorSelectionWidget(
-                          themeController: themeController,
-                        ),
-                        SizedBox(
-                          height: UiSizes.height_10,
-                        ),
-                        CustomCard(
-                          title: "Edit Profile",
-                          icon: FontAwesomeIcons.userPen,
-                          onTap: () {
-                            Navigator.pushNamed(context, AppRoutes.editProfile);
-                          },
-                        ),
-                        SizedBox(height: UiSizes.height_10),
-                        CustomCard(
-                          title: "Contribute to the project",
-                          icon: FontAwesomeIcons.github,
-                          onTap: () {
-                            Uri url = Uri.parse(githubRepoUrl);
-                            try {
-                              launchUrl(url);
-                            } catch (e) {
-                              log("Error launching URL: ${e.toString()}");
-                            }
-                          },
-                        ),
-                        SizedBox(height: UiSizes.height_10),
-                        CustomCard(
-                          title: "Terms and Conditions",
-                          icon: FontAwesomeIcons.fileInvoice,
-                          onTap: () {
-                            //TODO: Launch URL in webView
-                          },
-                        ),
-                        SizedBox(height: UiSizes.height_10),
-                        CustomCard(
-                          title: "Privacy Policy",
-                          icon: FontAwesomeIcons.shieldHalved,
-                          onTap: () {
-                            //TODO: Launch URL in webView
-                          },
-                        ),
-                        SizedBox(height: UiSizes.height_10),
-                        CustomCard(
-                          title: "Settings",
-                          icon: FontAwesomeIcons.gear,
-                          onTap: () =>
-                              Navigator.pushNamed(context, AppRoutes.settings),
-                        ),
-                        SizedBox(
-                          height: UiSizes.height_10,
-                        ),
-                      ],
+                              ),
+                            ),
+                          ),
+                          SizedBox(
+                            width: UiSizes.width_5,
+                          ),
+                          Expanded(
+                            child: SizedBox(
+                              height: 34,
+                              child: ElevatedButton(
+                                onPressed: () {
+                                  Get.toNamed(AppRoutes.settings);
+                                },
+                                child: const FittedBox(
+                                  child: Padding(
+                                    padding: EdgeInsets.all(8.0),
+                                    child: Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      children: [
+                                        Icon(Icons.settings),
+                                        SizedBox(
+                                          width: 10,
+                                        ),
+                                        Text("Settings"),
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
                     ),
-                  ),
-            emailVerifyController.isSending.value
-                ? BackdropFilter(
-                    filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
-                    child: Center(
-                        child: LoadingAnimationWidget.threeRotatingDots(
-                            color: Colors.blue, size: Get.pixelRatio * 20)),
-                  )
-                : const SizedBox(),
-          ])),
+                  ],
+                ),
+              ),
+              SizedBox(
+                height: UiSizes.height_20,
+              ),
+              Container(
+                padding: EdgeInsets.only(
+                  left: UiSizes.width_20,
+                ),
+                width: double.maxFinite,
+                child: Column(
+                  children: [
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Text(
+                        "Your Stories",
+                        style: TextStyle(
+                          fontSize: UiSizes.size_16,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                    ),
+                    SizedBox(
+                      height: UiSizes.height_5,
+                    ),
+                    SizedBox(
+                      height: UiSizes.height_200,
+                      child: ListView.builder(
+                        itemCount: yourStories.length,
+                        scrollDirection: Axis.horizontal,
+                        itemBuilder: (context, index) {
+                          return GestureDetector(
+                            onTap: () {
+                              Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                  builder: (context) => StoryScreen(
+                                    storyModel: yourStories[index],
+                                  ),
+                                ),
+                              );
+                            },
+                            child: Container(
+                              width: UiSizes.height_140,
+                              margin: EdgeInsets.only(
+                                right: UiSizes.width_10,
+                              ),
+                              child: Column(
+                                children: [
+                                  Container(
+                                    height: UiSizes.height_140,
+                                    width: UiSizes.height_140,
+                                    decoration: BoxDecoration(
+                                      color: Colors.black,
+                                      image: DecorationImage(
+                                        image: AssetImage(
+                                            yourStories[index].imgUrl),
+                                        fit: BoxFit.cover,
+                                      ),
+                                    ),
+                                  ),
+                                  SizedBox(
+                                    height: UiSizes.height_5,
+                                  ),
+                                  Text(
+                                    yourStories[index].title,
+                                    style: TextStyle(
+                                      fontSize: UiSizes.size_16,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                  Text(
+                                    yourStories[index].description,
+                                    style: TextStyle(
+                                      overflow: TextOverflow.ellipsis,
+                                      fontSize: UiSizes.size_12,
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          );
+                        },
+                      ),
+                    ),
+                    SizedBox(
+                      height: UiSizes.height_10,
+                    ),
+                    Align(
+                      alignment: Alignment.centerLeft,
+                      child: Text(
+                        "Continue with",
+                        style: TextStyle(
+                          fontSize: UiSizes.size_16,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                    ),
+                    SizedBox(
+                      height: UiSizes.height_5,
+                    ),
+                    SizedBox(
+                      height: UiSizes.height_200,
+                      child: ListView.builder(
+                        itemCount: continueWithStories.length,
+                        scrollDirection: Axis.horizontal,
+                        itemBuilder: (context, index) {
+                          return GestureDetector(
+                            onTap: () {
+                              Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                  builder: (context) => StoryScreen(
+                                    storyModel: continueWithStories[index],
+                                  ),
+                                ),
+                              );
+                            },
+                            child: Container(
+                              width: UiSizes.height_140,
+                              margin: EdgeInsets.only(
+                                right: UiSizes.width_10,
+                              ),
+                              child: Column(
+                                children: [
+                                  Container(
+                                    height: UiSizes.height_140,
+                                    width: UiSizes.height_140,
+                                    decoration: BoxDecoration(
+                                      color: Colors.black,
+                                      image: DecorationImage(
+                                        image: AssetImage(
+                                            continueWithStories[index].imgUrl),
+                                        fit: BoxFit.cover,
+                                      ),
+                                    ),
+                                  ),
+                                  SizedBox(
+                                    height: UiSizes.height_5,
+                                  ),
+                                  Text(
+                                    continueWithStories[index].title,
+                                    style: TextStyle(
+                                      fontSize: UiSizes.size_16,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                  Text(
+                                    continueWithStories[index].description,
+                                    style: TextStyle(
+                                      overflow: TextOverflow.ellipsis,
+                                      fontSize: UiSizes.size_12,
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          );
+                        },
+                      ),
+                    ),
+                  ],
+                ),
+              )
+            ],
+          ),
         ),
       ),
     );
