+++ lib/views/screens/settings_screen.dart
@@ -1,90 +1,116 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:get/get.dart';
+import 'package:get/get_rx/src/rx_typedefs/rx_typedefs.dart';
 import 'package:resonate/routes/app_routes.dart';
-import 'package:resonate/themes/theme_controller.dart';
+import 'package:resonate/utils/ui_sizes.dart';
 
-import '../../themes/choose_theme_bottom_sheet.dart';
+import '../../controllers/auth_state_controller.dart';
 
-class SettingsScreen extends StatefulWidget {
-  const SettingsScreen({super.key});
+class SettingsScreen extends StatelessWidget {
+  SettingsScreen({super.key});
 
-  @override
-  State<SettingsScreen> createState() => _SettingsScreenState();
-}
+  final authStateController =
+      Get.put<AuthStateController>(AuthStateController());
 
-class _SettingsScreenState extends State<SettingsScreen> {
-  final themeController = Get.find<ThemeController>();
+  final double padding = UiSizes.width_20;
 
-  late String subText;
+  Widget customTile({
+    required String str,
+    required Callback func,
+  }) {
+    return ListTile(
+      contentPadding: EdgeInsets.symmetric(
+        horizontal: padding,
+      ),
+      title: Text(str),
+      trailing: const Icon(Icons.arrow_forward_rounded),
+      onTap: func,
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
-    subText = themeController.loadTheme();
-    if (subText == 'systemDefault') {
-      subText = 'System Default';
-    } else if (subText == 'light') {
-      subText = 'Light';
-    } else {
-      subText = 'Dark';
+    Widget customDivider() {
+      return Divider(
+        height: UiSizes.height_30,
+        thickness: 10,
+        color: Theme.of(context).brightness == Brightness.light
+            ? Colors.black.withOpacity(0.04)
+            : Colors.white.withOpacity(0.04),
+      );
+    }
+
+    Widget titleText(String str) {
+      return Padding(
+        padding: EdgeInsets.only(
+          left: padding,
+          top: UiSizes.height_16,
+          bottom: UiSizes.height_10,
+        ),
+        child: Text(str,
+            style: TextStyle(
+              color: Theme.of(context).brightness == Brightness.light
+                  ? Colors.black54
+                  : Colors.white54,
+              fontWeight: FontWeight.bold,
+            )),
+      );
     }
 
     return Scaffold(
       appBar: AppBar(
-        title: const Text(
-          'Settings',
-        ),
+        title: const Text("Settings"),
       ),
-      body: Container(
-        padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 15),
-        child: ListView(
-          children: [
-            const Text('General settings'),
-            ListTile(
-              leading: const Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(FontAwesomeIcons.circleHalfStroke),
-                ],
-              ),
-              title: const Text('Theme'),
-              subtitle: Text(subText),
-              onTap: () {
-                showModalBottomSheet(
-                  shape: const RoundedRectangleBorder(
-                    borderRadius:
-                        BorderRadius.vertical(top: Radius.circular(24)),
-                  ),
-                  context: context,
-                  builder: (context) =>
-                      chooseThemeBottomSheet(context, (_) => setState(() {})),
-                );
-              },
+      body: ListView(
+        children: [
+          titleText("Account settings"),
+          customTile(
+            str: "Account",
+            func: () {
+              Get.toNamed(AppRoutes.userAccountScreen);
+            },
+          ),
+          customDivider(),
+          titleText("App settings"),
+          customTile(
+            str: "Themes",
+            func: () {
+              Get.toNamed(AppRoutes.themeScreen);
+            },
+          ),
+          customTile(
+            str: "About",
+            func: () {
+              Get.toNamed(AppRoutes.aboutApp);
+            },
+          ),
+          customDivider(),
+          titleText("Other"),
+          customTile(
+            str: "Contribute",
+            func: () {
+              Get.toNamed(AppRoutes.contributeScreen);
+            },
+          ),
+          customDivider(),
+          ListTile(
+            contentPadding: EdgeInsets.symmetric(
+              horizontal: padding,
             ),
-            const SizedBox(
-              height: 30,
-            ),
-            const Text('Account settings'),
-            Padding(
-              padding: const EdgeInsets.symmetric(vertical: 10),
-              child: OutlinedButton(
-                style: OutlinedButton.styleFrom(
-                    foregroundColor: Colors.redAccent,
-                    padding: const EdgeInsets.symmetric(vertical: 10),
-                    side: const BorderSide(color: Colors.redAccent)),
-                onPressed: () {
-                  Navigator.pushNamed(context, AppRoutes.deleteAccount);
-                },
-                child: const Text(
-                  'Delete Account',
-                  style: TextStyle(
-                    fontSize: 16,
-                  ),
-                ),
+            textColor: Colors.redAccent,
+            iconColor: Colors.redAccent,
+            title: const Text(
+              "Log out",
+              style: TextStyle(
+                fontWeight: FontWeight.bold,
               ),
             ),
-          ],
-        ),
+            trailing: const Icon(Icons.logout_rounded),
+            onTap: () async {
+              await authStateController.logout(context);
+            },
+          ),
+        ],
       ),
     );
   }
