+++ lib/views/screens/tabview_screen.dart
@@ -11,14 +11,15 @@ import 'package:resonate/controllers/rooms_controller.dart';
 import 'package:resonate/controllers/tabview_controller.dart';
 import 'package:resonate/themes/theme_controller.dart';
 import 'package:resonate/utils/ui_sizes.dart';
+import 'package:resonate/views/screens/create_room_screen.dart';
 import 'package:resonate/views/screens/discussions_screen.dart';
+import 'package:resonate/views/screens/home_screen.dart';
 import 'package:resonate/views/screens/home_screen.dart';
 import 'package:resonate/views/widgets/profile_avatar.dart';
 
 import '../../controllers/email_verify_controller.dart';
 import '../../utils/utils.dart';
 import '../widgets/pair_chat_dialog.dart';
-import 'create_room_screen.dart';
 
 class TabViewScreen extends StatelessWidget {
   final CreateRoomController createRoomController =
@@ -128,7 +129,6 @@ class TabViewScreen extends StatelessWidget {
             backgroundColor: Theme.of(context).brightness == Brightness.light
                 ? Colors.white
                 : const Color.fromRGBO(17, 17, 20, 1),
-            activeColor: themeController.primaryColor.value,
             inactiveColor: Theme.of(context).brightness == Brightness.light
                 ? Colors.black.withOpacity(0.3)
                 : Colors.white.withOpacity(0.3),
