+++ lib/views/widgets/welcome_screen_dialog.dart
@@ -54,30 +54,32 @@ Widget welcomeScreenDialog(BuildContext context) {
             SizedBox(
               height: UiSizes.height_20,
             ),
-            Container(
-              color: Colors.transparent,
-              height: UiSizes.height_40,
-              padding: EdgeInsets.symmetric(horizontal: UiSizes.width_20),
-
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    "Create new account",
-                    style: TextStyle(
-                      fontSize: UiSizes.size_15,
+            GestureDetector(
+              onTap: () {
+                Get.offNamed(AppRoutes.signup);
+              },
+              child: Container(
+                color: Colors.transparent,
+                height: UiSizes.height_40,
+                padding: EdgeInsets.symmetric(horizontal: UiSizes.width_20),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(
+                      "Create new account",
+                      style: TextStyle(
+                        fontSize: UiSizes.size_15,
+                      ),
                     ),
-                  ),
-                  const Icon(
-                    Icons.arrow_forward_rounded,
-                  )
-                ],
+                    const Icon(
+                      Icons.arrow_forward_rounded,
+                    )
+                  ],
+                ),
               ),
             ),
           ],
         ),
-
-
       ],
     ),
   );
